requirements.txt
pandas
numpy
scikit-learn
streamlit
joblib
matplotlib
groq
python-dotenv
reportlab
